refactor: update logo assets and improve document formatting in AMDProblem.pdf
