+++ frontend/src/components/ProblemPage/ProblemDesc.vue
@@ -151,7 +151,7 @@ import Background from '../ProblemPage/Background.vue'
 import NoticeBar from '../ProblemPage/NoticeBar.vue'
 import ProblemCardGroup from '../ProblemPage/ProblemCardGroup.vue'
 import ProblemCardGovernment from '../ProblemPage/ProblemCardGovernment.vue'
-import HsinchuLogo from '../../assets/Problems/Card/Hsinchulogo.svg'
+import HsinchuLogo from '../../assets/Problems/Logo/Government.webp'
 import BackgroundOne from '../ProblemPage/Bg1.vue'
 import BackgroundTwo from '../ProblemPage/Bg2.vue'
 import BackgroundThree from '../ProblemPage/Bg3.vue'
